fix: 解决miniexcel导出时会多出一列的问题，添加新的单元测试

+++ Excel/Collapsenav.Net.Tool.Excel.NPOI/Reader/NPOICellReader.cs
@@ -77,7 +77,7 @@ public class NPOICellReader : IExcelCellReader
     {
         get
         {
-            return new NPOICell(GetCell(GetRow(row), (int)col));
+            return new NPOICell(GetCell(GetRow(row), col));
         }
     }
     public IReadCell this[string field, int row] => new NPOICell(GetCell(GetRow(row), HeaderIndex[field]));
